refactor: refactor crm apps into a base class
