+++ .env.appStore.example
@@ -129,7 +129,7 @@ ZOHOCRM_CLIENT_SECRET=""
 
 
 #   - REVERT
-# Used for the Pipedrive integration (via/ Revert (https://revert.dev))
+# Used for the CRM integrations (via/ Revert (https://revert.dev))
 # @see https://github.com/calcom/cal.com/#obtaining-revert-api-keys
 REVERT_API_KEY=
 REVERT_PUBLIC_TOKEN=
